BC-233: Split out SecurityConfig tests.
